crawler : divar : set all field to save in db
